// src/lib/game-data-lib/updators/updator-player-planet-totals/updator-player-planet-totals-type.d.ts
export type UpdatorPlayerPlanetTotalsType = {
    influence: {
        avail: number;
        total: number;
    };
    resources: {
        avail: number;
        total: number;
    };
    techs: {
        blue: number;
        red: number;
        green: number;
        yellow: number;
    };
    traits: {
        cultural: number;
        hazardous: number;
        industrial: number;
    };
    legendary: number;
};